support intermediate flushes when encoding
due to the use of ready!(), whenever the underlying reader returns
Poll::Pending, it is transmitted directly to the caller, so there is no
flush until the entire data has been read.

This commit flushes compressed data when we get Poll::Pending, and stays
in the flushing state in case there is more data to send.